Fabric.js обновление логики #21 декомпозиция табов и фильтров
