+++ sorry-room/src/components/editor/Editor.vue
@@ -19,9 +19,7 @@ onMounted(() => {
     backgroundColor : "#fff",
   });
 
-  // const imageUrl = 'https://profil.mos.ru/images/banners/bottom/dit-banner.jpg';
   const imageUrl = 'https://fakestoreapi.com/img/71-3HjGNDUL._AC_SY879._SX._UX._SY._UY_.jpg';
-  // loadImageToSessionStorageAndAddToCanvas(imageUrl); // загрузка изображения и добавление в canvas
   fabric.Image.fromURL(imageUrl, (img) => {
     img.scaleToWidth(200);
     const ob = {layer: img, desc: 'img1'}
@@ -32,18 +30,7 @@ onMounted(() => {
   const rect = new fabric.Circle({ top: 10, left: 100, radius: 75, fill: "green", erasable: true });
   const ob = { layer: rect, desc: 'figure1'}
   editorStore.layers.push(ob);
-  // editorStore.layers.push(rect);
   editorStore.updateCanvas(fabricWrap.canvas); // обновляем canvas после добавления прямоугольника
-
-  // let font = new fabricWrap.canvas.Font('Ubuntu','/fonts/Ubuntu-Regular.ttf');
-  // font.addFace('/fonts/Ubuntu-Bold.ttf', 'bold');
-  // font.addFace('/fonts/Ubuntu-Italic.ttf', 'normal', 'italic');
-  // font.addFace('/fonts/Ubuntu-BoldItalic.ttf', 'bold', 'italic');
-  // fabricWrap.canvas.contextContainer.addFont(font)
-
-  // var fonts = ["Pacifico", "VT323", "Quicksand", "Inconsolata"];
-
-
 
   fabricWrap.canvas.on('selection:created', function(e) {
     const selectedObjects = e.selected;
@@ -70,34 +57,6 @@ onMounted(() => {
 
 });
 
-
-// function loadImageToSessionStorageAndAddToCanvas(url) {
-//   // Load image
-//   const image = new Image();
-//   image.crossOrigin = 'Anonymous'; // Устанавливаем настройку для кросс-доменных запросов
-//   image.onload = () => {
-//     // Создаем временный canvas для обработки изображения
-//     const tempCanvas = document.createElement('canvas');
-//     const tempContext = tempCanvas.getContext('2d');
-//     tempCanvas.width = 200;
-//     tempCanvas.height = (image.height / image.width) * tempCanvas.width;
-//     tempContext.drawImage(image, 0, 0, tempCanvas.width, tempCanvas.height);
-//
-//     // Получаем data URL изображения
-//     const dataURL = tempCanvas.toDataURL('image/png');
-//
-//     // Сохраняем data URL в sessionStorage
-//     sessionStorage.setItem('loadedImage', dataURL);
-//
-//     // Добавляем изображение на canvas
-//     fabric.Image.fromURL(dataURL, (img) => {
-//       img.scaleToWidth(200);
-//       fabricWrap.canvas.add(img);
-//     });
-//   };
-//   image.src = url;
-// }
-
   watch(() => editorStore.canvasOption, (state) => {
     // console.log(state)
     // fabricWrap.canvas.backgroundColor = editorStore.canvasOption.backgroundColor
@@ -107,7 +66,7 @@ onMounted(() => {
     fabricWrap.canvas.backgroundColor = state.backgroundColor
     // fabricWrap.canvas.setBackgroundImage(state.backgroundImg);
     fabricWrap.canvas.renderAll();
-  }, {deep: true}) //  гулбокий объект
+  }, {deep: true})
 
 
 
@@ -142,8 +101,8 @@ header {
   grid-template-columns: 2fr 1fr;
 }
 .container {
-  width: 800px; /* Фиксированная ширина контейнера */
-  height: 600px; /* Фиксированная высота контейнера */
-  overflow: auto; /* Добавление прокрутки в случае необходимости */
+  width: 800px;
+  height: 600px;
+  overflow: auto; /* прокрутка*/
 }
 </style>
