+++ sorry-room/src/components/editor/LayerFilters.vue
@@ -1,7 +1,8 @@
 <script setup lang="ts">
 import {useEditorStore} from "@/stores/editor";
-import {computed, onMounted, ref, watch} from "vue";
-import { fabric } from 'fabric-all-modules';
+import {computed, ref} from "vue";
+import ImagesFilter from "@/components/editor/filters/imagesFilter.vue";
+import TextFilter from "@/components/editor/filters/textFilter.vue";
 
 const props = defineProps({
   fabricWrap: Object,
@@ -10,19 +11,6 @@ const props = defineProps({
 const editorStore = useEditorStore();
 
 const open = ref([])
-
-const filters = ['grayscale',  'sepia', 'brownie', 'vintage', 'kodachrome', 'technicolor', 'polaroid']
-const fonts = ['Arial', 'Verdana', 'Tahoma', 'Times New Roman', 'Courier New', 'Georgia']
-const fontsWeight = ['bold', 'normal', '400', '600', '800']
-const fontsStyle = ['normal', 'italic', 'oblique']
-
-// const activeFilters = ref(false)
-// const activeText = ref(false)
-function getAct() {
-  // const obj = props.fabricWrap.canvas.getActiveObject();
-  // console.log(obj)
-  // editorStore.filterBlockedLayer(props.fabricWrap.canvas)
-}
 
 const isDisabled = computed(() => {
   if (editorStore.selectedLayerIndex !== -1)
@@ -40,97 +28,6 @@ const FilterBlockedLayer = computed({
     editorStore.layers[editorStore.selectedLayerIndex].layer.set('erasable', value)
   }
 });
-
-const appliedFilters = computed({
-  get: () => {
-    if (editorStore.selectedLayerIndex !== -1) {
-    const activeFilters = editorStore.layers[editorStore.selectedLayerIndex].layer.filters
-    let arrActiveNameFilter = []
-    // console.log(activeFilters)
-    if (activeFilters !== undefined && activeFilters.length > 0 ) {
-      activeFilters.forEach((element) => arrActiveNameFilter.push(element.type.toLowerCase()))
-    }
-    return arrActiveNameFilter
-    } else return []
-  },
-  set: (value) => {
-    editorStore.layers[editorStore.selectedLayerIndex].layer.filters = []
-    // if (editorStore.layers[editorStore.selectedLayerIndex].layer.text === undefined) {
-    if (value.length > 0) {
-      value.forEach(filterName => {
-        switch (filterName) {
-          case "brownie":
-            editorStore.layers[editorStore.selectedLayerIndex].layer.filters.push(new fabric.Image.filters.Brownie())
-            break;
-          case "sepia":
-            editorStore.layers[editorStore.selectedLayerIndex].layer.filters.push(new fabric.Image.filters.Sepia());
-            break;
-          case "vintage":
-            editorStore.layers[editorStore.selectedLayerIndex].layer.filters.push(new fabric.Image.filters.Vintage())
-            break;
-          case "technicolor":
-            editorStore.layers[editorStore.selectedLayerIndex].layer.filters.push(new fabric.Image.filters.Technicolor())
-            break;
-          case "polaroid":
-            editorStore.layers[editorStore.selectedLayerIndex].layer.filters.push(new fabric.Image.filters.Polaroid())
-            break;
-          case "kodachrome":
-            editorStore.layers[editorStore.selectedLayerIndex].layer.filters.push(new fabric.Image.filters.Kodachrome())
-            break;
-          case "grayscale":
-            editorStore.layers[editorStore.selectedLayerIndex].layer.filters.push(new fabric.Image.filters.Grayscale())
-            break;
-        }
-      })
-    }
-    editorStore.layers[editorStore.selectedLayerIndex].layer.applyFilters()
-  }
-});
-
-const filterTextFont = computed({
-  get: () => {
-    if (editorStore.selectedLayerIndex !== -1) {
-    return editorStore.layers[editorStore.selectedLayerIndex].layer.get('fontFamily')
-    } else return false
-  },
-  set: (value) => {
-    editorStore.layers[editorStore.selectedLayerIndex].layer.set('fontFamily', value)
-  }
-});
-
-const filterTextColor = computed({
-  get: () => {
-    if (editorStore.selectedLayerIndex !== -1) {
-    return editorStore.layers[editorStore.selectedLayerIndex].layer.get('fill')
-    } else return false
-  },
-  set: (value) => {
-    editorStore.layers[editorStore.selectedLayerIndex].layer.set('fill', value)
-  }
-});
-
-const filterTextFontWeight = computed({
-  get: () => {
-    if (editorStore.selectedLayerIndex !== -1) {
-    return editorStore.layers[editorStore.selectedLayerIndex].layer.get('fontWeight')
-    } else return false
-  },
-  set: (value) => {
-    editorStore.layers[editorStore.selectedLayerIndex].layer.set('fontWeight', value)
-  }
-});
-
-const filterTextFontStyle = computed({
-  get: () => {
-    if (editorStore.selectedLayerIndex !== -1) {
-    return editorStore.layers[editorStore.selectedLayerIndex].layer.get('fontStyle')
-    } else return false
-  },
-  set: (value) => {
-    editorStore.layers[editorStore.selectedLayerIndex].layer.set('fontStyle', value)
-  }
-});
-
 </script>
 
 <template>
@@ -142,13 +39,8 @@ const filterTextFontStyle = computed({
   >
 
     <v-toolbar color="purple">
-<!--      <v-btn icon="mdi-menu" @click="getAct()"></v-btn>-->
-
       <v-toolbar-title>Settings Layer</v-toolbar-title>
-
       <v-spacer></v-spacer>
-
-<!--      <v-btn icon="mdi-magnify"></v-btn>-->
     </v-toolbar>
 
     <v-list lines="three" select-strategy="classic" v-model:opened="open">
@@ -163,48 +55,9 @@ const filterTextFontStyle = computed({
         <v-list-item-title>Blocked layer</v-list-item-title>
       </v-list-item>
 
-      <v-list-group value="Filters">
-        <template v-slot:activator="{ props }">
-          <v-list-item
-            v-bind="props"
-            prepend-icon="mdi-filter-settings-outline"
-            title="Filters"
-          ></v-list-item>
-        </template>
-          <v-select
-            :items="filters"
-            v-model="appliedFilters"
-            label="filters"
-            multiple
-            chips
-          ></v-select>
-      </v-list-group>
+      <images-filter></images-filter>
 
-      <v-list-group value="Text">
-        <template v-slot:activator="{ props }">
-          <v-list-item
-            v-bind="props"
-            prepend-icon="mdi-script-text-key-outline"
-            title="Text filter"
-          ></v-list-item>
-        </template>
-        <v-select
-          label="Font"
-          v-model="filterTextFont"
-          :items="fonts"
-        ></v-select>
-        <v-select
-          label="Font Style"
-          v-model="filterTextFontStyle"
-          :items="fontsStyle"
-        ></v-select>
-        <v-select
-          label="Font Weight"
-          v-model="filterTextFontWeight"
-          :items="fontsWeight"
-        ></v-select>
-        <v-text-field label="Color" v-model="filterTextColor"></v-text-field>
-      </v-list-group>
+      <text-filter></text-filter>
     </v-list>
   </v-card>
 </section>
